Add flexibility for django project directory structure
Previously we required the `manage.py` script to be located in the
project root directory.  Now we allow the project to place the django
project (i.e., directory containing `manage.py`) in a subdirectory of
the project directory.  We still require it to be no more than 1
subdirectory from the project root though.